fix bug

+++ app/assets/stylesheets/favorite.js
@@ -1,32 +1,40 @@
 document.addEventListener("DOMContentLoaded",function(){
     const token = document.head.querySelector('meta[name="csrf-token"]').content
+    document.querySelectorAll(".favorite-link").forEach(function(link){
+        const action = link.classList[1].split("-")[0];
+        if(action == "delete"){
+            link.setAttribute("data-fav-id",link.href.split("/").slice(-1)[0])
+        }
+    })
     document.querySelectorAll(".favorite-link").forEach(function(addLink){
         addLink.addEventListener("click",function(e){
             e.preventDefault();
             const blogId = addLink.classList[1].split("-")[1];
             const firstElement = addLink.classList[1].split("-")[0];
             if(firstElement == "delete"){
-                const favId = addLink.getAttribute("fav-id")
-                fetch("/favorites/" + favId,{
-                    method: "DELETE",
-                    headers:  {
-                        "Content-Type": "application/json",
-                        "Accept": "application/json"
-                    },
-                    body : JSON.stringify({ authenticity_token: token, entity: { name: "new name" } })
-                }).then(function(res){
-                    addLink.classList.remove(addLink.classList[1]);
-                    addLink.parentNode.lastChild.textContent = (Number)(addLink.parentNode.lastChild.textContent) - 1
-                    if(addLink.querySelector("i").textContent === "star"){
-                        addLink.querySelector("i").textContent = "star_border";
-                    }else{
-                        addLink.querySelector("i").textContent = "star";
-                    }
-                    addLink.classList.add(`post-${blogId}`)
-                    addLink.removeAttribute("fav-id")
-                }).catch(function(){
-                    alert("Failed to remove from favorites")
-                });
+                const favId = addLink.getAttribute("data-fav-id")
+                if(favId != null){
+                    fetch("/favorites/" + favId,{
+                        method: "DELETE",
+                        headers:  {
+                            "Content-Type": "application/json",
+                            "Accept": "application/json"
+                        },
+                        body : JSON.stringify({ authenticity_token: token, entity: { name: "new name" } })
+                    }).then(function(res){
+                        addLink.classList.remove(addLink.classList[1]);
+                        addLink.parentNode.lastChild.textContent = (Number)(addLink.parentNode.lastChild.textContent) - 1
+                        if(addLink.querySelector("i").textContent === "star"){
+                            addLink.querySelector("i").textContent = "star_border";
+                        }else{
+                            addLink.querySelector("i").textContent = "star";
+                        }
+                        addLink.classList.add(`post-${blogId}`)
+                        addLink.removeAttribute("data-fav-id")
+                    }).catch(function(){
+                        alert("Failed to remove from favorites")
+                    });
+                }
             }else if(firstElement == 'post'){
                 fetch("/favorites?blog_id=" + blogId,{
                     method: "POST",
@@ -46,7 +54,7 @@ document.addEventListener("DOMContentLoaded",function(){
                         addLink.querySelector("i").textContent = "star";
                     }
                     addLink.classList.add(`delete-${blogId}`)
-                    addLink.setAttribute("fav-id",data.id)
+                    addLink.setAttribute("data-fav-id",data.id)
                 }).catch(function(){
                     alert("Failed to add to favorites")
                 });
